Extract `cvd/fetch/build_strings.{cc,h}`
Bug: b/424021019
Test: bazel test '//...' --test_summary=terse
